+++ CHANGELOG.md
@@ -11,6 +11,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Starship weapon firing arcs.
 - When hovering over a weapon with a firing arc, a template will be rendered in the canvas, representing that fire arc.
 - Starship Modification DataModel.
+- Starship tokens can now track passive skill values, power dice, and fuel amount.
+- (Dis)Advantage flags for starship skills.
+- Enrichment lookup (`[[/skill skillName]]`) for starship skills.
 
 ### Changes
 
@@ -31,6 +34,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Quantity of starship modifiations will be used for the modifications counter.
 - Deployments and their features should now work properly on the NPC sheet.
 - Uses for starship modifications should now appear in the inventory.
+- "Apply damage with resistances" context button should once again work.
 
 ## [2.4.0.2.7.0] - 2033-11-30